fixed CI check issues
