+++ weevent-governance/src/main/java/com/webank/weevent/governance/GovernanceApplication.java
@@ -1,22 +1,13 @@
 package com.webank.weevent.governance;
 
-import java.io.IOException;
 import java.io.InterruptedIOException;
-import java.net.HttpURLConnection;
-import java.net.InetAddress;
-import java.net.Socket;
 import java.net.UnknownHostException;
 import java.security.KeyManagementException;
 import java.security.NoSuchAlgorithmException;
 import java.security.cert.X509Certificate;
 
-import javax.net.ssl.HostnameVerifier;
-import javax.net.ssl.HttpsURLConnection;
 import javax.net.ssl.SSLContext;
 import javax.net.ssl.SSLException;
-import javax.net.ssl.SSLSession;
-import javax.net.ssl.SSLSocket;
-import javax.net.ssl.SSLSocketFactory;
 import javax.net.ssl.TrustManager;
 import javax.net.ssl.X509TrustManager;
 
@@ -25,12 +16,6 @@ import com.webank.weevent.governance.filter.ForwardProcessorFilter;
 import com.webank.weevent.governance.filter.ForwardWebaseFilter;
 import com.webank.weevent.governance.filter.UserAuthFilter;
 import com.webank.weevent.governance.filter.XssFilter;
-import com.webank.weevent.governance.handler.JsonAccessDeniedHandler;
-import com.webank.weevent.governance.handler.JsonAuthenticationEntryPoint;
-import com.webank.weevent.governance.handler.JsonLogoutSuccessHandler;
-import com.webank.weevent.governance.handler.LoginFailHandler;
-import com.webank.weevent.governance.properties.ConstantProperties;
-import com.webank.weevent.governance.service.AccountDetailsService;
 import com.webank.weevent.governance.utils.H2ServerUtil;
 
 import lombok.extern.slf4j.Slf4j;
@@ -50,7 +35,6 @@ import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClients;
 import org.apache.http.impl.conn.PoolingHttpClientConnectionManager;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -60,15 +44,8 @@ import org.springframework.boot.web.servlet.ServletRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Scope;
 import org.springframework.http.client.ClientHttpRequestFactory;
-import org.springframework.http.client.SimpleClientHttpRequestFactory;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.builders.WebSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 import org.springframework.web.context.WebApplicationContext;
 import org.springframework.web.servlet.DispatcherServlet;
@@ -81,7 +58,6 @@ import org.springframework.web.servlet.DispatcherServlet;
  */
 @Slf4j
 @SpringBootApplication
-@EnableWebSecurity
 @EnableTransactionManagement
 public class GovernanceApplication {
 
@@ -119,28 +95,6 @@ public class GovernanceApplication {
     @Autowired
     private ForwardProcessorFilter forwardProcessorFilter;
 
-    //browerSecurity
-    @Autowired
-    private AccountDetailsService userDetailService;
-
-    @Qualifier(value = "loginSuccessHandler")
-    @Autowired
-    private AuthenticationSuccessHandler loginSuccessHandler;
-
-    @Qualifier(value = "loginFailHandler")
-    @Autowired
-    private LoginFailHandler loginfailHandler;
-
-    @Autowired
-    private JsonAuthenticationEntryPoint jsonAuthenticationEntryPoint;
-
-    @Autowired
-    private JsonAccessDeniedHandler jsonAccessDeniedHandler;
-
-    @Autowired
-    private JsonLogoutSuccessHandler jsonLogoutSuccessHandler;
-
-
     private PoolingHttpClientConnectionManager cm;
 
 
@@ -163,56 +117,8 @@ public class GovernanceApplication {
     }
 
     @Bean
-    public WebSecurityConfigurerAdapter initBrowerSecurityConfig() {
-        WebSecurityConfigurerAdapter webSecurityConfigurerAdapter = new WebSecurityConfigurerAdapter() {
-            @Override
-            protected void configure(HttpSecurity http) throws Exception {
-                http.exceptionHandling().accessDeniedHandler(jsonAccessDeniedHandler);
-
-                http.formLogin() // define user login page
-                        .loginPage("/user/require")
-                        .loginProcessingUrl("/user/login")
-                        .usernameParameter("username")
-                        .passwordParameter("password")
-                        .permitAll()
-                        .successHandler(loginSuccessHandler) // if login success
-                        .failureHandler(loginfailHandler) // if login fail
-                        .and()
-                        .authorizeRequests()
-                        .antMatchers("/user/**", "/", "/static/**", "/weevent-governance/user/**")
-                        .permitAll()
-                        .anyRequest()
-                        .authenticated()
-                        .and()
-                        .csrf()
-                        .disable()
-                        .httpBasic()
-                        .authenticationEntryPoint(jsonAuthenticationEntryPoint)
-                        .and()
-                        .logout()
-                        .logoutUrl("/user/logout")
-                        .deleteCookies(ConstantProperties.COOKIE_JSESSIONID, ConstantProperties.COOKIE_MGR_ACCOUNT)
-                        .logoutSuccessHandler(jsonLogoutSuccessHandler)
-                        .permitAll();
-            }
-
-            @Override
-            protected void configure(AuthenticationManagerBuilder auth) throws Exception {
-                auth.userDetailsService(userDetailService).passwordEncoder(new BCryptPasswordEncoder());
-            }
-
-            @Override
-            public void configure(WebSecurity web) throws Exception {
-                web.ignoring().antMatchers("/static/**");
-            }
-
-        };
-        return webSecurityConfigurerAdapter;
-    }
-
-    @Bean
     public ClientHttpRequestFactory httpsClientRequestFactory() {
-        SimpleClientHttpRequestFactory factory = initHttpsFactory();
+        HttpsClientRequestFactory factory = new HttpsClientRequestFactory();
         factory.setReadTimeout(readTimeout);// ms
         factory.setConnectTimeout(connectTimeOut);// ms
         return factory;
@@ -368,119 +274,6 @@ public class GovernanceApplication {
 
         public void checkClientTrusted(X509Certificate[] certs, String authType) {
             // don't check
-        }
-    }
-
-
-    @Bean
-    public SimpleClientHttpRequestFactory initHttpsFactory() {
-        SimpleClientHttpRequestFactory simpleClientHttpRequestFactory = new SimpleClientHttpRequestFactory() {
-            @Override
-            protected void prepareConnection(HttpURLConnection connection, String httpMethod) throws IOException {
-                try {
-                    if (!(connection instanceof HttpsURLConnection)) {
-                        throw new RuntimeException("An instance of HttpsURLConnection is expected");
-                    }
-
-                    HttpsURLConnection httpsConnection = (HttpsURLConnection) connection;
-
-                    TrustManager[] trustAllCerts = new TrustManager[]{new X509TrustManager() {
-                        @Override
-                        public X509Certificate[] getAcceptedIssuers() {
-                            return null;
-                        }
-
-                        @Override
-                        public void checkClientTrusted(X509Certificate[] certs, String authType) {
-                        }
-
-                        @Override
-                        public void checkServerTrusted(X509Certificate[] certs, String authType) {
-                        }
-
-                    }};
-
-                    SSLContext sslContext = SSLContext.getInstance("TLS");
-                    sslContext.init(null, trustAllCerts, new java.security.SecureRandom());
-                    httpsConnection.setSSLSocketFactory(new MyCustomSSLSocketFactory(sslContext.getSocketFactory()));
-
-                    httpsConnection.setHostnameVerifier(new HostnameVerifier() {
-                        @Override
-                        public boolean verify(String s, SSLSession sslSession) {
-                            return true;
-                        }
-                    });
-
-                    super.prepareConnection(httpsConnection, httpMethod);
-                } catch (Exception e) {
-                    log.error(e.getMessage());
-                }
-            }
-
-
-        };
-        return simpleClientHttpRequestFactory;
-    }
-
-    /**
-     *  We need to invoke sslSocket.setEnabledProtocols(new String[] {"SSLv3"});
-     */
-    class MyCustomSSLSocketFactory extends SSLSocketFactory {
-        private final SSLSocketFactory delegate;
-
-        public MyCustomSSLSocketFactory(SSLSocketFactory delegate) {
-            this.delegate = delegate;
-        }
-
-        @Override
-        public String[] getDefaultCipherSuites() {
-            return delegate.getDefaultCipherSuites();
-        }
-
-        @Override
-        public String[] getSupportedCipherSuites() {
-            return delegate.getSupportedCipherSuites();
-        }
-
-        @Override
-        public Socket createSocket(final Socket socket, final String host, final int port, final boolean autoClose)
-                throws IOException {
-            final Socket underlyingSocket = delegate.createSocket(socket, host, port, autoClose);
-            return overrideProtocol(underlyingSocket);
-        }
-
-        @Override
-        public Socket createSocket(final String host, final int port) throws IOException {
-            final Socket underlyingSocket = delegate.createSocket(host, port);
-            return overrideProtocol(underlyingSocket);
-        }
-
-        @Override
-        public Socket createSocket(final String host, final int port, final InetAddress localAddress,
-                                   final int localPort) throws IOException {
-            final Socket underlyingSocket = delegate.createSocket(host, port, localAddress, localPort);
-            return overrideProtocol(underlyingSocket);
-        }
-
-        @Override
-        public Socket createSocket(final InetAddress host, final int port) throws IOException {
-            final Socket underlyingSocket = delegate.createSocket(host, port);
-            return overrideProtocol(underlyingSocket);
-        }
-
-        @Override
-        public Socket createSocket(final InetAddress host, final int port, final InetAddress localAddress,
-                                   final int localPort) throws IOException {
-            final Socket underlyingSocket = delegate.createSocket(host, port, localAddress, localPort);
-            return overrideProtocol(underlyingSocket);
-        }
-
-        private Socket overrideProtocol(final Socket socket) {
-            if (!(socket instanceof SSLSocket)) {
-                throw new RuntimeException("An instance of SSLSocket is expected");
-            }
-            ((SSLSocket) socket).setEnabledProtocols(new String[]{"TLSv1"});
-            return socket;
         }
     }
 }
